+++ privacyidea/lib/smsprovider/HttpSMSProvider.py
@@ -48,27 +48,12 @@ from privacyidea.lib.smsprovider.SMSProvider import (ISMSProvider, SMSError)
 from privacyidea.lib import _
 import requests
 from urllib.parse import urlparse
-import re
 import logging
 
 log = logging.getLogger(__name__)
 
 
 class HttpSMSProvider(ISMSProvider):
-
-    @staticmethod
-    def _mangle_phone(phone, config):
-        regexp = config.get("REGEXP")
-        if regexp:
-            try:
-                m = re.match("^/(.*)/(.*)/$", regexp)
-                if m:
-                    phone = re.sub(m.group(1), m.group(2), phone)
-            except re.error:
-                log.warning("Can not mangle phone number. "
-                            "Please check your REGEXP: {0!s}".format(regexp))
-
-        return phone
 
     def submit_message(self, phone, message):
         """
